move layout to layouts folder
